requirements.txt
Pillow==9.4.0
pytesseract==0.3.10
stdl==0.4.1
